+++ src/main/java/com/spring/SpeedAuction/Controller/TestController.java
@@ -6,6 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+
+// Api test
 @RestController
 @RequestMapping("/api/test")
 public class TestController {
