+++ src/main/java/com/spring/SpeedAuction/Controller/UserController.java
@@ -3,7 +3,7 @@ package com.spring.SpeedAuction.Controller;
 import com.spring.SpeedAuction.Models.UserModels;
 import com.spring.SpeedAuction.dto.FavouriteDTO;
 import com.spring.SpeedAuction.dto.UserResponsDTO;
-import com.spring.SpeedAuction.security.Services.UserServices;
+import com.spring.SpeedAuction.Security.Services.UserServices;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
