+++ src/main/java/com/spring/SpeedAuction/Enums/EColor.java
@@ -1,4 +1,4 @@
-package com.spring.SpeedAuction.enums;
+package com.spring.SpeedAuction.Enums;
 
 public enum EColor {
     BLACK, RED, WHITE, BLUE, GREEN, YELLOW, ORANGE, PURPLE, BROWN, PINK, GRAY, SILVER, OTHER
