+++ src/main/java/com/spring/SpeedAuction/Enums/ECondition.java
@@ -1,4 +1,4 @@
-package com.spring.SpeedAuction.enums;
+package com.spring.SpeedAuction.Enums;
 
 public enum ECondition {
     GOOD,DECENT,BAD
