+++ src/main/java/com/spring/SpeedAuction/Enums/ERole.java
@@ -1,4 +1,4 @@
-package com.spring.SpeedAuction.enums;
+package com.spring.SpeedAuction.Enums;
 
 public enum ERole {
     ROLE_USER, ROLE_MODERATOR, ROLE_ADMIN
