+++ src/main/java/com/spring/SpeedAuction/Payload/response/UserInfoResponse.java
@@ -1,37 +1,19 @@
-package com.spring.SpeedAuction.payload.response;
+package com.spring.SpeedAuction.Payload.response;
 
 import java.util.List;
 
-public class JwtResponse {
-    private String token;
-    private String type = "Bearer";
+public class UserInfoResponse {
+
     private String id;
     private String username;
     private String email;
     private List<String> roles;
 
-    public JwtResponse(String accessToken, String id, String username,String email, List<String>roles){
-        this.token = accessToken;
+    public UserInfoResponse(String id, String username, String email, List<String> roles) {
         this.id = id;
         this.username = username;
         this.email = email;
         this.roles = roles;
-    }
-
-    public String getAccessToken() {
-        return token;
-    }
-
-    public void setAccessToken(String accessToken) {
-        this.token = accessToken;
-    }
-
-    public String getTokenType() {
-        return type;
-    }
-
-    public void setTokenType(String tokenType ) {
-        this.type = tokenType;
     }
 
     public String getId() {
@@ -60,5 +42,9 @@ public class JwtResponse {
 
     public List<String> getRoles() {
         return roles;
+    }
+
+    public void setRoles(List<String> roles) {
+        this.roles = roles;
     }
 }
