+++ src/main/java/com/spring/SpeedAuction/Repository/BidsRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
 
-public interface BidsModelsRepository extends MongoRepository<BidsModels,String> {
+public interface BidsRepository extends MongoRepository<BidsModels,String> {
     List<BidsModels> findBidsModelsByBidderIdOrderByTimeBiddedDesc(String bidderId);
     List<BidsModels> findBidsModelsByAuctionIdOrderByAmountDesc(String auctionId);
     BidsModels findFirstByAuctionIdOrderByAmountDesc(String auctionId);
