+++ src/main/java/com/spring/SpeedAuction/Security/Services/UserDetailImpl.java
@@ -1,4 +1,4 @@
-package com.spring.SpeedAuction.security.Services;
+package com.spring.SpeedAuction.Security.Services;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.spring.SpeedAuction.Models.UserModels;
 import org.springframework.security.core.GrantedAuthority;
