+++ src/main/java/com/spring/SpeedAuction/Security/Services/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.spring.SpeedAuction.security.Services;
+package com.spring.SpeedAuction.Security.Services;
 
 import com.spring.SpeedAuction.Models.UserModels;
 import com.spring.SpeedAuction.Repository.UserRepository;
