+++ src/main/java/com/spring/SpeedAuction/Security/Services/UserServices.java
@@ -1,4 +1,4 @@
-package com.spring.SpeedAuction.security.Services;
+package com.spring.SpeedAuction.Security.Services;
 
 import com.spring.SpeedAuction.Models.AuctionModels;
 import com.spring.SpeedAuction.Models.UserModels;
